+++ app/src/index.ts
@@ -14,11 +14,25 @@ import type {
   SwitchViewEventData,
   ErrorEventData,
 } from './types/types';
+import type {
+  TimepickerOptions,
+  ClockOptions,
+  UIOptions,
+  LabelsOptions,
+  BehaviorOptions,
+  CallbacksOptions,
+} from './types/options';
 import { EventEmitter } from './utils/EventEmitter';
 
 export {
   TimepickerUI,
   OptionTypes,
+  TimepickerOptions,
+  ClockOptions,
+  UIOptions,
+  LabelsOptions,
+  BehaviorOptions,
+  CallbacksOptions,
   OpenEventData,
   CancelEventData,
   ConfirmEventData,